Combine everything into a single file.
This works around a compiler bug.
